+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- set-hosts: new command to add custom etc/hosts entries to a pot
 - set-env: new command to add environment variable to a pot
 - network-type private-bridge: add a new network layout, to provide private bridges for a group of pots
 - create-private-bridge: new subcommand to define and create a private-bridge
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - POT_EXTRA_EXTIF: add addition network interfaces support
 
 ### Changed
+- start: overwrite /etc/hosts of a pot, adding all pots on the same bridge and custom entries added via set-hosts
 - flavorable commands: extend support to set-cmd and set-env
 - pot-rdr anchor: the name of the anchor is now a truncated pot name (the last 54 characters)